enterprise_ai_orchestration/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.24
pydantic>=2.0
pyyaml>=6.0
scikit-learn>=1.1
numpy>=1.23
